+++ jsonjsdb/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - add: escapeHtml to prevent XSS attacks when rendering data in HTML context
 - fix: tables type and and checkIntegrity method in Jsonjsdb class
+- refactor: use idSuffix variable to improve flexibility and maintainability
 
 ## 0.7.1 (2025-09-29)